Make kafka encoder/decoder bolts shared
This bolt's can/should be used as base for any encoder/decored bolts in
any topology.
