+++ services/wfm/src/main/java/org/openkilda/wfm/share/bolt/KafkaEncoder.java
@@ -1,0 +1,97 @@
+/* Copyright 2018 Telstra Open Source
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package org.openkilda.wfm.share.bolt;
+
+import org.openkilda.messaging.Message;
+import org.openkilda.messaging.MessageData;
+import org.openkilda.messaging.Utils;
+import org.openkilda.messaging.command.CommandData;
+import org.openkilda.messaging.command.CommandMessage;
+import org.openkilda.messaging.info.InfoData;
+import org.openkilda.messaging.info.InfoMessage;
+import org.openkilda.wfm.AbstractBolt;
+import org.openkilda.wfm.CommandContext;
+import org.openkilda.wfm.error.AbstractException;
+import org.openkilda.wfm.error.JsonEncodeException;
+import org.openkilda.wfm.error.PipelineException;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import org.apache.storm.kafka.bolt.mapper.FieldNameBasedTupleToKafkaMapper;
+import org.apache.storm.topology.OutputFieldsDeclarer;
+import org.apache.storm.tuple.Fields;
+import org.apache.storm.tuple.Tuple;
+import org.apache.storm.tuple.Values;
+
+public abstract class KafkaEncoder extends AbstractBolt {
+    public static final String FIELD_ID_PAYLOAD = "payload";
+
+    public static final Fields STREAM_FIELDS = new Fields(
+            FieldNameBasedTupleToKafkaMapper.BOLT_KEY, FieldNameBasedTupleToKafkaMapper.BOLT_MESSAGE);
+
+    @Override
+    protected void handleInput(Tuple input) throws AbstractException {
+        MessageData payload = pullPayload(input);
+        try {
+            Message message = wrap(pullContext(input), payload);
+            String json = encode(message);
+
+            getOutput().emit(input, new Values(null, json));
+        } catch (IllegalArgumentException e) {
+            log.error(e.getMessage());
+            unhandledInput(input);
+        }
+    }
+
+    protected MessageData pullPayload(Tuple input) throws PipelineException {
+        return pullValue(input, FIELD_ID_PAYLOAD, MessageData.class);
+    }
+
+    protected Message wrap(CommandContext commandContext, MessageData payload) {
+        Message message = null;
+        if (payload instanceof CommandData) {
+            message = wrapCommand(commandContext, (CommandData) payload);
+        } else if (payload instanceof InfoData) {
+            message = wrapInfo(commandContext, (InfoData) payload);
+        } else {
+            throw new IllegalArgumentException(String.format("There is not rule to build envelope for: %s", payload));
+        }
+
+        return message;
+    }
+
+    private CommandMessage wrapCommand(CommandContext commandContext, CommandData payload) {
+        return new CommandMessage(payload, System.currentTimeMillis(), commandContext.getCorrelationId());
+    }
+
+    private InfoMessage wrapInfo(CommandContext commandContext, InfoData payload) {
+        return new InfoMessage(payload, System.currentTimeMillis(), commandContext.getCorrelationId());
+    }
+
+    private String encode(Message message) throws JsonEncodeException {
+        String value;
+        try {
+            value = Utils.MAPPER.writeValueAsString(message);
+        } catch (JsonProcessingException e) {
+            throw new JsonEncodeException(message, e);
+        }
+        return value;
+    }
+
+    @Override
+    public void declareOutputFields(OutputFieldsDeclarer outputManager) {
+        outputManager.declare(STREAM_FIELDS);
+    }
+}
