+++ services/wfm/src/main/java/org/openkilda/wfm/topology/ping/bolt/FlowStatusEncoder.java
@@ -15,63 +15,22 @@
 
 package org.openkilda.wfm.topology.ping.bolt;
 
-import org.openkilda.messaging.Utils;
-import org.openkilda.messaging.info.InfoMessage;
+import org.openkilda.messaging.MessageData;
 import org.openkilda.messaging.info.flow.FlowPingReport;
 import org.openkilda.messaging.model.PingReport;
-import org.openkilda.wfm.AbstractBolt;
-import org.openkilda.wfm.CommandContext;
-import org.openkilda.wfm.error.AbstractException;
-import org.openkilda.wfm.error.JsonEncodeException;
 import org.openkilda.wfm.error.PipelineException;
+import org.openkilda.wfm.share.bolt.KafkaEncoder;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import org.apache.storm.kafka.bolt.mapper.FieldNameBasedTupleToKafkaMapper;
-import org.apache.storm.topology.OutputFieldsDeclarer;
-import org.apache.storm.tuple.Fields;
 import org.apache.storm.tuple.Tuple;
-import org.apache.storm.tuple.Values;
 
-public class FlowStatusEncoder extends AbstractBolt {
+public class FlowStatusEncoder extends KafkaEncoder {
     public static final String BOLT_ID = ComponentId.FLOW_STATUS_ENCODER.toString();
 
     public static final String FIELD_ID_PING_REPORT = "flow_status";
 
-    public static final Fields STREAM_FIELDS = new Fields(
-            FieldNameBasedTupleToKafkaMapper.BOLT_KEY, FieldNameBasedTupleToKafkaMapper.BOLT_MESSAGE);
-
     @Override
-    protected void handleInput(Tuple input) throws AbstractException {
-        PingReport report = pullPingReport(input);
-        InfoMessage message = wrap(input, report);
-        String json = encode(message);
-
-        Values output = new Values(null, json);
-        getOutput().emit(input, output);
-    }
-
-    private InfoMessage wrap(Tuple input, PingReport pingReport) throws PipelineException {
-        FlowPingReport payload = new FlowPingReport(pingReport);
-        CommandContext commandContext = pullContext(input);
-        return new InfoMessage(payload, System.currentTimeMillis(), commandContext.getCorrelationId());
-    }
-
-    private String encode(InfoMessage message) throws JsonEncodeException {
-        String json;
-        try {
-            json = Utils.MAPPER.writeValueAsString(message);
-        } catch (JsonProcessingException e) {
-            throw new JsonEncodeException(message, e);
-        }
-        return json;
-    }
-
-    private PingReport pullPingReport(Tuple input) throws PipelineException {
-        return pullValue(input, FIELD_ID_PING_REPORT, PingReport.class);
-    }
-
-    @Override
-    public void declareOutputFields(OutputFieldsDeclarer outputManager) {
-        outputManager.declare(STREAM_FIELDS);
+    protected MessageData pullPayload(Tuple input) throws PipelineException {
+        PingReport pingReport = pullValue(input, FIELD_ID_PING_REPORT, PingReport.class);
+        return new FlowPingReport(pingReport);
     }
 }
