+++ services/wfm/src/main/java/org/openkilda/wfm/topology/ping/bolt/SpeakerEncoder.java
@@ -15,58 +15,8 @@
 
 package org.openkilda.wfm.topology.ping.bolt;
 
-import org.openkilda.messaging.Utils;
-import org.openkilda.messaging.command.CommandData;
-import org.openkilda.messaging.command.CommandMessage;
-import org.openkilda.wfm.CommandContext;
-import org.openkilda.wfm.error.AbstractException;
-import org.openkilda.wfm.error.JsonEncodeException;
-import org.openkilda.wfm.error.PipelineException;
+import org.openkilda.wfm.share.bolt.KafkaEncoder;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import org.apache.storm.kafka.bolt.mapper.FieldNameBasedTupleToKafkaMapper;
-import org.apache.storm.topology.OutputFieldsDeclarer;
-import org.apache.storm.tuple.Fields;
-import org.apache.storm.tuple.Tuple;
-import org.apache.storm.tuple.Values;
-
-public class SpeakerEncoder extends Abstract {
+public class SpeakerEncoder extends KafkaEncoder {
     public static final String BOLT_ID = ComponentId.SPEAKER_ENCODER.toString();
-
-    public static final String FIELD_ID_PAYLOAD = "payload";
-
-    public static final Fields STREAM_FIELDS = new Fields(
-            FieldNameBasedTupleToKafkaMapper.BOLT_KEY, FieldNameBasedTupleToKafkaMapper.BOLT_MESSAGE);
-
-    @Override
-    protected void handleInput(Tuple input) throws AbstractException {
-        CommandData payload = pullPayload(input);
-        CommandMessage message = wrap(pullContext(input), payload);
-        String json = encode(message);
-
-        getOutput().emit(new Values(null, json));
-    }
-
-    private CommandData pullPayload(Tuple input) throws PipelineException {
-        return pullValue(input, FIELD_ID_PAYLOAD, CommandData.class);
-    }
-
-    private CommandMessage wrap(CommandContext commandContext, CommandData payload) {
-        return new CommandMessage(payload, System.currentTimeMillis(), commandContext.getCorrelationId());
-    }
-
-    private String encode(CommandMessage message) throws JsonEncodeException {
-        String value;
-        try {
-            value = Utils.MAPPER.writeValueAsString(message);
-        } catch (JsonProcessingException e) {
-            throw new JsonEncodeException(message, e);
-        }
-        return value;
-    }
-
-    @Override
-    public void declareOutputFields(OutputFieldsDeclarer outputManager) {
-        outputManager.declare(STREAM_FIELDS);
-    }
 }
